refactor(trip-viewer-overlay): Replace TripViewerOverlay with its counterpart from OTP-UI.
