+++ lib/components/map/default-map.js
@@ -6,15 +6,16 @@ import {
   carRentalQuery,
   vehicleRentalQuery
 } from '../../actions/api'
+
 import BaseMap from './base-map'
 import EndpointsOverlay from './endpoints-overlay'
 import ParkAndRideOverlay from './park-and-ride-overlay'
+import RouteViewerOverlay from './route-viewer-overlay'
 import StopsOverlay from './stops-overlay'
 import StopViewerOverlay from './stop-viewer-overlay'
 import TileOverlay from './tile-overlay'
 import TransitiveOverlay from './transitive-overlay'
-import TripViewerOverlay from './trip-viewer-overlay'
-import RouteViewerOverlay from './route-viewer-overlay'
+import TripViewerOverlay from './connected-trip-viewer-overlay'
 import VehicleRentalOverlay from './vehicle-rental-overlay'
 import ZipcarOverlay from './zipcar-overlay'
