+++ package.json
@@ -33,6 +33,7 @@
     "@opentripplanner/geocoder": "^0.0.18",
     "@opentripplanner/location-field": "^0.0.18",
     "@opentripplanner/trip-details": "^0.0.18",
+    "@opentripplanner/trip-viewer-overlay": "^0.0.18",
     "bootstrap": "^3.3.7",
     "clone": "^2.1.0",
     "connected-react-router": "^6.5.2",
